Implemented ability to build TnFOX as a no-GUI build which contains almost no FOX at all

+++ include/qptrvector.h
@@ -111,7 +111,7 @@ public:
 	//! Extends the list to \em no items if not that already
 	bool extend(uint no)
 	{
-		uint cs=std::vector<type *>::size();
+		uint cs=(uint) std::vector<type *>::size();
 		if(cs<no)
 		{
 			std::vector<type *>::resize(no);
